+++ packages/client/src/game/data/colors.ts
@@ -1,0 +1,18 @@
+export enum Color {
+  Yellow = 'yellow',
+  Lime = 'lime',
+  Black = 'black',
+  Grey = '#636363',
+  LightGrey = 'lightgrey',
+  Orange = 'red',
+  Red = 'red',
+  White = 'white',
+  Transparent = 'transparent',
+  LightCyan = 'lightcyan',
+  Aqua = 'aqua',
+  Brown = 'brown',
+  Green = 'green',
+  Blue = 'blue',
+  AliceBlue = 'aliceblue',
+  Pink = '#483D8B',
+}
