+++ packages/client/src/game/entities/Entity.ts
@@ -1,3 +1,4 @@
+import { Color } from '../data/colors';
 import {
   Animations,
   AnimationSettings,
@@ -29,8 +30,8 @@ export abstract class Entity extends EventEmitter<EntityEvent> {
   flying = false;
   crossable = false;
   hittable = true;
-  // В этом файле нельзя подключать Color из Constants т.к. образуются циклические зависимости и тесты падают.
-  color = 'grey';
+
+  color: Color | string = Color.Grey;
   shouldBeDestroyed = false;
   /** Значение true делает танк неуязвимым для снарядов. */
   invincible = false;
