+++ packages/client/src/game/entities/TankEnemy.ts
@@ -1,4 +1,5 @@
-import { Color, spriteCoordinates } from '../data/constants';
+import { Color } from '../data/colors';
+import { spriteCoordinates } from '../data/constants';
 import { type EntityDynamicSettings, Direction, EntityEvent, Speed } from '../typings';
 import { rand } from '../utils';
 import { Tank } from './Tank';
@@ -55,6 +56,7 @@ export class TankEnemy extends Tank {
 
   getRandomAction() {
     const actions = [...new Array(1).fill('turn'), ...new Array(5).fill('move')];
+
     return actions[Math.floor(Math.random() * actions.length)];
   }
