+++ packages/client/src/game/entities/Terrain.ts
@@ -1,4 +1,5 @@
-import { Color, spriteCoordinates } from '../data/constants';
+import { Color } from '../data/colors';
+import { spriteCoordinates } from '../data/constants';
 import { EntityEvent, EntitySettings, TerrainVariant } from '../typings';
 import { Entity } from './';
